+++ h/models/annotation.py
@@ -10,8 +10,7 @@ from sqlalchemy.ext.hybrid import hybrid_property
 from sqlalchemy.ext.mutable import MutableDict, MutableList
 
 from h.db import Base, types
-from h.util import uri
-from memex import markdown
+from h.util import markdown, uri
 
 
 class Annotation(Base):
