+++ requirements.in
@@ -3,6 +3,7 @@ SQLAlchemy >= 1.1.4
 alembic
 backports.functools_lru_cache
 bcrypt
+bleach >= 2.0
 celery == 3.1.25  # Pin to latest 3.1.x to ensure forwards-compat with 4.x
 certifi
 cffi
@@ -18,6 +19,7 @@ itsdangerous
 jsonpointer == 1.0
 jsonschema
 kombu <3.1 # Pinned because of celery pin
+mistune
 newrelic
 passlib
 psycogreen
